+++ packages/common/src/core/schemas/internal/metrics/ProjectUiSchemaMetrics.ts
@@ -62,8 +62,8 @@ export const buildUiSchema = async (
             options: {
               control: "hub-field-input-tile-select",
               rules: [
-                undefined,
-                undefined,
+                [],
+                [],
                 [
                   {
                     effect: UiSchemaRuleEffects.SHOW,