+++ packages/common/test/core/schemas/internal/metrics/ProjectUiSchemaMetrics.test.ts
@@ -51,8 +51,8 @@ describe("ProjectUiSchemaMetrics", () => {
                 options: {
                   control: "hub-field-input-tile-select",
                   rules: [
-                    undefined,
-                    undefined,
+                    [],
+                    [],
                     [
                       {
                         effect: UiSchemaRuleEffects.SHOW,